Move exporter strategies to dedicated package.
