+++ refinedstorage2-network-api/src/main/java/com/refinedmods/refinedstorage2/api/network/node/exporter/strategy/ExporterTransferStrategy.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage2.api.network.node.exporter;
+package com.refinedmods.refinedstorage2.api.network.node.exporter.strategy;
 
 import com.refinedmods.refinedstorage2.api.network.Network;
 import com.refinedmods.refinedstorage2.api.storage.Actor;
