+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/network/node/exporter/ExporterTransferStrategyFactory.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage2.platform.api.network.node.exporter;
 
-import com.refinedmods.refinedstorage2.api.network.node.exporter.ExporterTransferStrategy;
+import com.refinedmods.refinedstorage2.api.network.node.exporter.strategy.ExporterTransferStrategy;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
